+++ src/components/PageFooter/PageFooter.tsx
@@ -1,10 +1,9 @@
 import { FC } from 'react';
 import { Footer, Link, FooterItem } from '@taskany/bricks';
-import { textColor } from '@taskany/colors';
 
 import { ModalEvent, dispatchModalEvent } from '../../utils/dispatchModal';
-import { Light } from '../Light';
 
+import s from './PageFooter.module.css';
 import { tr } from './PageFooter.i18n';
 
 export const PageFooter: FC = () => {
@@ -17,9 +16,9 @@ export const PageFooter: FC = () => {
     ];
     return (
         <Footer>
-            <Light color={textColor}>
-                <FooterItem onClick={dispatchModalEvent(ModalEvent.FeedbackCreateModal)}>{tr('Feedback')}</FooterItem>
-            </Light>
+            <FooterItem onClick={dispatchModalEvent(ModalEvent.FeedbackCreateModal)} className={s.FooterItem}>
+                {tr('Feedback')}
+            </FooterItem>
 
             {menuItems.map(({ title, url }) => (
                 <Link key={url} href={url} inline>
