+++ src/components/ReactionsButton.tsx
@@ -1,7 +1,6 @@
 import React, { useCallback } from 'react';
-import styled from 'styled-components';
 import { Button } from '@taskany/bricks/harmony';
-import { IconMoodTongueOutline, iconSizesMap } from '@taskany/icons';
+import { BaseIcon, IconMoodTongueOutline } from '@taskany/icons';
 
 interface ReactionsButtonProps {
     emoji?: string;
@@ -10,11 +9,6 @@ interface ReactionsButtonProps {
     onClick?: (emoji?: string) => void;
 }
 
-const StyledEmoji = styled.span`
-    font-size: ${iconSizesMap.s}px;
-`;
-
-// eslint-disable-next-line react/display-name
 export const ReactionsButton = React.memo(
     React.forwardRef<HTMLButtonElement, ReactionsButtonProps>(({ emoji, count, onClick }, ref) => {
         const onButtonClick = useCallback(
@@ -28,13 +22,7 @@ export const ReactionsButton = React.memo(
             <Button
                 ref={ref}
                 text={count && count > 1 ? String(count) : undefined}
-                iconLeft={
-                    emoji ? (
-                        <StyledEmoji dangerouslySetInnerHTML={{ __html: emoji }} />
-                    ) : (
-                        <IconMoodTongueOutline size="xs" />
-                    )
-                }
+                iconLeft={emoji ? <BaseIcon size="s" value={() => emoji} /> : <IconMoodTongueOutline size="xs" />}
                 onClick={onButtonClick(emoji)}
             />
         );
